refactor auth system

+++ pkg/store/tokens.go
@@ -7,6 +7,7 @@ import (
 	"github.com/jmoiron/sqlx"
 )
 
+// Define the scoped of the tokens used in the application.
 const (
 	ScopeActivation      = "activation"
 	ScopeRecoverMainKeys = "recover-keys"
@@ -21,16 +22,15 @@ type Token struct {
 	CreatedAt time.Time `db:"created_at" json:"-"`
 }
 
+// The store abstraction to manipulate tokens into the database. It holds a DB
+// connection pool. Only the hashed version of the token are saved into the db.
 type TokenStore struct {
-	db *sqlx.DB
+	DB *sqlx.DB
 }
 
-func NewTokensStore(db *sqlx.DB) TokenStore {
-	return TokenStore{
-		db: db,
-	}
-}
-
+// Creates a new token with the given scope and ttl (time-to-live) and saves the hash into the
+// database. The plain text version of the token is returned and not viewable/recoverable
+// anymore.
 func (m *TokenStore) New(userID int64, ttl time.Duration, scope string) (Token, error) {
 	plainToken, tokenHash, err := generateToken()
 	if err != nil {
@@ -51,32 +51,30 @@ func (m *TokenStore) New(userID int64, ttl time.Duration, scope string) (Token, 
 	return token, nil
 }
 
+// Insert a new token into the database. The plain text version of the token
+// is not saved.
 func (m *TokenStore) Insert(token Token) error {
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	err := m.db.GetContext(ctx, &token, `
+	return m.DB.GetContext(ctx, &token, `
 		INSERT INTO tokens (hash, user_id, expiry, scope) 
 		VALUES ($1, $2, $3, $4)
 		RETURNING created_at
 	`, token.Hash, token.UserID, token.Expiry, token.Scope)
-
-	return err
 }
 
+// Delete all tokens with the given scope for the specified user.
 func (m *TokenStore) DeleteAllForUser(scope string, userID int64) error {
 
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	res, err := m.db.ExecContext(ctx, `
-		DELETE FROM tokens WHERE scope = $1 AND user_id = $2
-	`, scope, userID)
+	res, err := m.DB.ExecContext(ctx, `DELETE FROM tokens WHERE scope = $1 AND user_id = $2`, scope, userID)
 	if err != nil {
 		return err
 	}
-
 	n, err := res.RowsAffected()
 	if err != nil {
 		return err
